+++ source/project.h
@@ -42,7 +42,7 @@ class Project
 {
 public:
 
-	Project(const std::string& pFilename,size_t pNumThreads,bool pVerboseOutput,bool pRebuild,int pTruncateOutput);
+	Project(const std::string& pFilename,size_t pNumThreads,bool pVerboseOutput,bool pRebuild,size_t pTruncateOutput);
 	~Project();
 
 	operator bool ()const{return mOk;}
@@ -68,7 +68,7 @@ private:
 	const size_t mNumThreads;
 	const bool mVerboseOutput;
 	const bool mRebuild;
-	const int mTruncateOutput;
+	const size_t mTruncateOutput;
 	
 	// This project file, fully pathed.
 	const std::string mPathedProjectFilename;
